apps/backend/requirements.txt
fastapi==0.115.6
uvicorn==0.30.6
python-jose==3.3.0
passlib[bcrypt]==1.7.4
